+++ lib/src/state_machine/mp_th2_file_edit_state_machine/mixins/mp_th2_file_edit_state_move_canvas_mixin.dart
@@ -20,11 +20,10 @@ mixin MPTH2FileEditStateMoveCanvasMixin on MPTH2FileEditState {
 
     switch (event.logicalKey) {
       case LogicalKeyboardKey.keyA:
-        if ((isCtrlPressed || isMetaPressed) &&
+        if (!isCtrlPressed &&
+            !isMetaPressed &&
             !isAltPressed &&
             !isShiftPressed) {
-          selectionController.selectAllElements();
-        } else if (!isAltPressed && !isShiftPressed) {
           th2FileEditController.stateController.setState(
             MPTH2FileEditStateType.addArea,
           );