+++ lib/src/state_machine/mp_th2_file_edit_state_machine/mp_th2_file_edit_state_edit_single_line.dart
@@ -74,6 +74,13 @@ class MPTH2FileEditStateEditSingleLine extends MPTH2FileEditState
           elementEditController.applyRemoveSelectedLineSegments();
           keyProcessed = true;
         }
+      case LogicalKeyboardKey.keyA:
+        if ((isCtrlPressed || isMetaPressed) &&
+            !isAltPressed &&
+            !isShiftPressed) {
+          onSelectAll();
+          keyProcessed = true;
+        }
       case LogicalKeyboardKey.keyL:
         keyProcessed = onKeyLDownEvent(event);
     }