+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Infrastructure maintenance:
   * flutter pub upgrade --major-versions.
   * Created onDeselectAll() and onSelectAll() state and actuator methods.
+  * Migrate select and desselect all elements to use onDeselectAll() and onSelectAll().
 
 ## 0.2.26 - 2025-12-11 - The São Paulo release
 * New features: